Avoid warnings when the file is not a file or not an expercted one
git-svn-id: https://cpc.regardscitoyens.org/svn/cpc/trunk/project@3413 9954c6d0-a799-495f-bc23-bbe1ca3f12d5

+++ lib/task/loadTagCommissionFromDossierTask.class.php
@@ -16,16 +16,18 @@ class loadTagCommissionFromDossierTask extends sfBaseTask {
     if (is_dir($dir)) {
       if ($dh = opendir($dir)) {
         while (($file = readdir($dh)) != false) {
-          foreach(file($dir.$file) as $line) {
-	    if (preg_match('/([^;]+);(\d+)/', $line, $matches)) {
-	      print $matches[1].' => '.$matches[2]."\n";
-	      foreach ( Doctrine::getTable('Intervention')->findBySource($matches[1]) as $i) {
-		$i->addTag('loi:numero='.$matches[2]);
-		$i->save();
+	  if (preg_match('/.asp/', $file)) {
+	      foreach(file($dir.$file) as $line) {
+		if (preg_match('/([^;]+);(\d+)/', $line, $matches)) {
+		  print $matches[1].' => '.$matches[2]."\n";
+		  foreach ( Doctrine::getTable('Intervention')->findBySource($matches[1]) as $i) {
+		    $i->addTag('loi:numero='.$matches[2]);
+		    $i->save();
+		  }
+		}
 	      }
-	    }
+	      unlink($dir.$file);
 	  }
-	  unlink($dir.$file);
 	}
       }
     }
